+++ app/src/Screens/RegisterProject/PickNameStep.svelte
